+++ modules/algebra/src/Sphere3D.cpp
@@ -17,11 +17,11 @@ IMPALGEBRA_BEGIN_NAMESPACE
 Sphere3D::Sphere3D(const Vector3D& center,double radius):center_(center),
                                                  radius_(radius){
 }
-Float Sphere3D::get_volume() const {
+double Sphere3D::get_volume() const {
   return PI * (4.0 / 3.0) * std::pow(radius_, 3.0);
 }
 
-Float Sphere3D::get_surface_area() const {
+double Sphere3D::get_surface_area() const {
   return PI * 4.0 * square(radius_);
 }
 bool Sphere3D::get_contains(const Vector3D &p) const {
